Avoid running browser polyfills inside tests
