+++ extension/sitelist.js
@@ -1,7 +1,5 @@
 // @flow
 /* @jsx h */
-/* global window */
-import browser from 'webextension-polyfill';
 
 // Utils to manage the whitelist and blacklist of URLs
 // where Hyper Gwent is enabled
@@ -63,25 +61,6 @@ function whitelist(
 ) {
     removeUrlRule(disabledSites, url);
     addUrlRule(enabledSites, url);
-}
-
-/*
- * Returns the URL of the active tab
- */
-export async function getCurrentUrl(): Promise<URL> {
-    if (!browser.tabs) {
-        // Executing in a normal page context
-        return window.location;
-    }
-    // Executing in the extension popup menu
-
-    // The only one tab returned should be the active one
-    const [activeTab] = await browser.tabs.query({
-        active: true,
-        currentWindow: true
-    });
-
-    return new URL(activeTab.url);
 }
 
 // -----------------------------------
